+++ test/build-chain-flow.test.js
@@ -7,8 +7,8 @@ jest.mock("../src/lib/build-chain-flow-helper");
 const {
   readWorkflowInformation,
   checkoutParentsAndGetWorkflowInformation
-} = require("../src/lib/workflow-informaton-reader");
-jest.mock("../src/lib/workflow-informaton-reader");
+} = require("../src/lib/workflow-information/reader");
+jest.mock("../src/lib/workflow-information/reader");
 const {
   run: runUploadArtifactsMock
 } = require("../src/lib/artifacts/upload-artifacts");
@@ -139,7 +139,12 @@ test("start no parent dependencies archive artifacts", async () => {
     buildCommandsUpstream: ["upstream 1"],
     buildCommandsDownstream: ["downstream 1"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifact1",
       dependencies: "none"
     }
@@ -163,11 +168,19 @@ test("start no parent dependencies archive artifacts", async () => {
   await start(context);
   // Assert
   expect(runUploadArtifactsMock).toHaveBeenCalledTimes(1);
-  expect(runUploadArtifactsMock).toHaveBeenCalledWith({
-    name: "artifact1",
-    path: "whateverpath",
-    dependencies: "none"
-  });
+  expect(runUploadArtifactsMock).toHaveBeenCalledWith(
+    {
+      name: "artifact1",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
+      dependencies: "none"
+    },
+    ["success"]
+  );
   expect(checkoutProject).toHaveBeenCalledTimes(1);
   expect(checkoutProject).toHaveBeenCalledWith(
     context,
@@ -413,7 +426,12 @@ test("start with parent dependencies with archive artifacts with path", async ()
     buildCommandsUpstream: ["command-child-upstream"],
     buildCommandsDownstream: ["command-child-downstream"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifactChild",
       dependencies: ["projectXParent"]
     }
@@ -431,7 +449,12 @@ test("start with parent dependencies with archive artifacts with path", async ()
     buildCommands: ["command-parent"],
     buildCommandsDownstream: ["command-parent-dowstream"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifactParent",
       dependencies: "none"
     }
@@ -475,16 +498,32 @@ test("start with parent dependencies with archive artifacts with path", async ()
   );
 
   expect(runUploadArtifactsMock).toHaveBeenCalledTimes(2);
-  expect(runUploadArtifactsMock).toHaveBeenCalledWith({
-    path: "whateverpath",
-    name: "artifactChild",
-    dependencies: ["projectXParent"]
-  });
-  expect(runUploadArtifactsMock).toHaveBeenCalledWith({
-    path: "whateverpath",
-    name: "artifactParent",
-    dependencies: "none"
-  });
+  expect(runUploadArtifactsMock).toHaveBeenCalledWith(
+    {
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
+      name: "artifactChild",
+      dependencies: ["projectXParent"]
+    },
+    ["success"]
+  );
+  expect(runUploadArtifactsMock).toHaveBeenCalledWith(
+    {
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
+      name: "artifactParent",
+      dependencies: "none"
+    },
+    ["success"]
+  );
 });
 
 test("start with parent dependencies with archive artifacts with path dependencies all", async () => {
@@ -516,7 +555,12 @@ test("start with parent dependencies with archive artifacts with path dependenci
     buildCommandsUpstream: ["command-child-upstream"],
     buildCommandsDownstream: ["command-child-downstream"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifactChild",
       dependencies: "all"
     }
@@ -534,7 +578,12 @@ test("start with parent dependencies with archive artifacts with path dependenci
     buildCommands: ["command-parent"],
     buildCommandsDownstream: ["command-parent-dowstream"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifactParent",
       dependencies: "none"
     }
@@ -578,16 +627,32 @@ test("start with parent dependencies with archive artifacts with path dependenci
   );
 
   expect(runUploadArtifactsMock).toHaveBeenCalledTimes(2);
-  expect(runUploadArtifactsMock).toHaveBeenCalledWith({
-    path: "whateverpath",
-    name: "artifactChild",
-    dependencies: "all"
-  });
-  expect(runUploadArtifactsMock).toHaveBeenCalledWith({
-    path: "whateverpath",
-    name: "artifactParent",
-    dependencies: "none"
-  });
+  expect(runUploadArtifactsMock).toHaveBeenCalledWith(
+    {
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
+      name: "artifactChild",
+      dependencies: "all"
+    },
+    ["success"]
+  );
+  expect(runUploadArtifactsMock).toHaveBeenCalledWith(
+    {
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
+      name: "artifactParent",
+      dependencies: "none"
+    },
+    ["success"]
+  );
 });
 
 test("start with parent dependencies with archive artifacts with path dependencies not matching", async () => {
@@ -619,7 +684,12 @@ test("start with parent dependencies with archive artifacts with path dependenci
     buildCommandsUpstream: ["command-child-upstream"],
     buildCommandsDownstream: ["command-child-downstream"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifactChild",
       dependencies: ["projectXParentX"]
     }
@@ -637,7 +707,12 @@ test("start with parent dependencies with archive artifacts with path dependenci
     buildCommands: ["command-parent"],
     buildCommandsDownstream: ["command-parent-dowstream"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifactParent",
       dependencies: "none"
     }
@@ -681,11 +756,19 @@ test("start with parent dependencies with archive artifacts with path dependenci
   );
 
   expect(runUploadArtifactsMock).toHaveBeenCalledTimes(1);
-  expect(runUploadArtifactsMock).toHaveBeenCalledWith({
-    path: "whateverpath",
-    name: "artifactChild",
-    dependencies: ["projectXParentX"]
-  });
+  expect(runUploadArtifactsMock).toHaveBeenCalledWith(
+    {
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
+      name: "artifactChild",
+      dependencies: ["projectXParentX"]
+    },
+    ["success"]
+  );
   expect(printCheckoutInformation).toHaveBeenCalledTimes(1);
   expect(printCheckoutInformation).toHaveBeenCalledWith({});
 });
@@ -736,7 +819,12 @@ test("start with parent dependencies with archive artifacts one of them without 
     buildCommands: ["command-parent"],
     buildCommandsDownstream: ["command-parent-dowstream"],
     archiveArtifacts: {
-      path: "whateverpath",
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
       name: "artifactParent",
       dependencies: "none"
     }
@@ -780,11 +868,19 @@ test("start with parent dependencies with archive artifacts one of them without 
   );
 
   expect(runUploadArtifactsMock).toHaveBeenCalledTimes(1);
-  expect(runUploadArtifactsMock).toHaveBeenCalledWith({
-    path: "whateverpath",
-    name: "artifactParent",
-    dependencies: "none"
-  });
+  expect(runUploadArtifactsMock).toHaveBeenCalledWith(
+    {
+      paths: [
+        {
+          path: "whateverpath",
+          on: "success"
+        }
+      ],
+      name: "artifactParent",
+      dependencies: "none"
+    },
+    ["success"]
+  );
   expect(printCheckoutInformation).toHaveBeenCalledTimes(1);
   expect(printCheckoutInformation).toHaveBeenCalledWith({});
 });